backend/requirements.txt
Flask == 3.0.3
icecream == 2.1.3
mysql-connector-python == 9.0.0
flask-session == 0.8.0
requests == 2.32.5
flask-cors==4.0.1
